Adds the skeleton of the new structure. Folder waise.
