+++ Sources/Cells/TabbyCell.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class TabbyCell {
+  
+}
