+++ Sources/Structs/TabbyBarItem.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct TabbyBarItem {
+
+  var controller: UIViewController
+  var image: UIImage
+}
